refactor: move the tmux binding record schema out of the binding store
Extract the record types, validators, cloning, identity matching, and
expiry helpers into tmuxBindingRecords.ts (931 lines). The binding
store keeps the serialized store class and the filesystem IO helpers,
imports the record layer back, and re-exports the original public
surface. Verbatim move; no behavior change.
